Start of a game API: state retrieval
It's not REST but it's JSON. With this API you can get the full state of
the game, as a person in the normal game would be able to observe.

It's hoped that the client code will one day use this information to
display something meaningful to the player, perhaps resembling the game
of Uno.