Start of a game API
- start/end a game
- draw/place a card
  -> the server doesn't know when a player is +2 or +4 affected which
  could be abused, a column has to be added to the users table.

Also, black cards (plusfour, joker) have to be colored by the client
before sending. The idea is that when a player chooses to place a black
card, the UI asks the desired new color, prefixes it to the card name
like "red-plusfour" and makes the query to the server.
That means the "color" column of the games table is now useless.